+++ src/schema/graph_schema.rs
@@ -25,9 +25,7 @@ impl SchemaDatabaseOperation for GraphSchema {
         silent: bool,
     ) -> Result<Option<Self::PoolType>, ClientError> {
         log::debug!("Creating Graph {}", &self.0.name);
-        let duplicate = serde_json::to_string(self).unwrap();
-        let duplicate: Self = serde_json::from_str(&duplicate).unwrap();
-        let graph = duplicate.into();
+        let graph = self.clone().into();
         Self::handle_pool_result(database.create_graph(graph, true).await, silent)
     }